chore: refactor test fixtures
- Refactor test fixtures to use fixturify-project in favor of scenario-tester.
  - Decouple file and project fixtures from scenarios.
- Remove duplicate scenario test.
- Refactor migration-graph.test.ts and migration-strategy.test.ts to use new proejct fixtures.
