+++ packages/migration-graph/package.json
@@ -34,6 +34,7 @@
     "@types/minimatch": "^5.1.2",
     "@types/rimraf": "^3.0.2",
     "fixturify": "^2",
+    "fixturify-project": "^5.0.3",
     "lodash.merge": "^4.6.2",
     "micromatch": "^4.0.4",
     "rimraf": "^3.0.2",
