+++ packages/migration-graph/tests/migration-graph.test.ts
@@ -1,10 +1,18 @@
-import { PreparedApp } from 'scenario-tester';
-import { beforeAll, describe, expect, test } from 'vitest';
+import merge from 'lodash.merge';
+import tmp from 'tmp';
+import { describe, expect, test } from 'vitest';
 
 import { GraphNode, UniqueGraphNode } from '../src';
 import { buildMigrationGraph, DetectedSource } from '../src/migration-graph';
 import { getLibrarySimple } from './fixtures/library';
-import { getEmberAddon, getEmberApp, setup } from './fixtures/scenarios';
+import {
+  getEmberAddonProject,
+  getEmberAppProject,
+  getEmberAppWithInRepoAddonProject,
+  getEmberAppWithInRepoEngineProject,
+  setupProject,
+} from './fixtures/project';
+tmp.setGracefulCleanup();
 
 function flatten(arr: GraphNode<UniqueGraphNode>[]): string[] {
   return Array.from(arr).map((n) => n.content.key);
@@ -13,8 +21,8 @@ function flatten(arr: GraphNode<UniqueGraphNode>[]): string[] {
 describe('migration-graph', () => {
   describe('package', () => {
     test('simple', () => {
-      const rootDir = getLibrarySimple();
-      const m = buildMigrationGraph(rootDir);
+      const baseDir = getLibrarySimple();
+      const m = buildMigrationGraph(baseDir);
 
       expect(m.graph.hasNode('my-package')).toBe(true);
       expect(m.sourceType).toBe(DetectedSource.Library);
@@ -27,12 +35,10 @@ describe('migration-graph', () => {
   });
 
   describe('ember', () => {
-    beforeAll(() => {
-      setup();
-    });
     test('app', async () => {
-      const app: PreparedApp = await getEmberApp('app');
-      const m = buildMigrationGraph(app.dir);
+      const project = await setupProject(getEmberAppProject());
+
+      const m = buildMigrationGraph(project.baseDir);
       expect(m.sourceType, 'should detect an EmberApp').toBe(DetectedSource.EmberApp);
       const orderedPackages = m.graph.topSort();
       expect(flatten(orderedPackages)).toStrictEqual(['app-template']);
@@ -43,8 +49,10 @@ describe('migration-graph', () => {
       ]);
     });
     test('app-with-in-repo-addon', async () => {
-      const app: PreparedApp = await getEmberApp('appWithInRepoAddon');
-      const m = buildMigrationGraph(app.dir);
+      const project = await setupProject(getEmberAppWithInRepoAddonProject());
+
+      const m = buildMigrationGraph(project.baseDir);
+
       expect(m.sourceType, 'should detect an EmberApp').toBe(DetectedSource.EmberApp);
       const orderedPackages = m.graph.topSort();
       expect(flatten(orderedPackages)).toStrictEqual(['some-addon', 'app-template']);
@@ -60,8 +68,38 @@ describe('migration-graph', () => {
       ).toStrictEqual(['app/app.js', 'app/components/salutation.js', 'app/router.js']);
     });
     test('app-with-in-repo-engine', async () => {
-      const app: PreparedApp = await getEmberApp('appWithInRepoEngine');
-      const m = buildMigrationGraph(app.dir);
+      const { name: tmpDir } = tmp.dirSync();
+      const project = getEmberAppWithInRepoEngineProject();
+
+      // TODO refactor requirePackageMain() to parse the index.js file first, if no name found result to require.
+
+      // This is a workaround because for this test we don't want to have to do a `install`
+      // on the fixture
+      //
+      // We are mocking require(`ember-engines/lib/engine-addon`) for now because
+      // in `migration-graph-ember` there is a method called `requirePackageMain` which
+      // calls `require()` on the `index.js` of the addon to determine the addon-name.
+      // This test fails unless we mock that module.
+      //
+
+      project.files = merge(project.files, {
+        node_modules: {
+          'ember-engines': {
+            lib: {
+              'engine-addon': `module.exports = {
+                extend: function(config) {
+                  return config;
+                }
+              }`,
+            },
+          },
+        },
+      });
+
+      project.baseDir = tmpDir;
+      await project.write();
+
+      const m = buildMigrationGraph(project.baseDir);
       expect(m.sourceType, 'should detect an EmberApp').toBe(DetectedSource.EmberApp);
       const orderedPackages = m.graph.topSort();
       expect(flatten(orderedPackages)).toStrictEqual(['some-engine', 'app-template']);
@@ -76,11 +114,12 @@ describe('migration-graph', () => {
       ).toStrictEqual(['app/app.js', 'app/components/salutation.js', 'app/router.js']);
     });
     test('addon', async () => {
-      const app: PreparedApp = await getEmberAddon('addon');
-      const m = buildMigrationGraph(app.dir);
+      const project = await setupProject(getEmberAddonProject());
+
+      const m = buildMigrationGraph(project.baseDir);
       expect(m.sourceType, 'should detect an EmberAddon').toBe(DetectedSource.EmberAddon);
       const orderedPackages = m.graph.topSort();
-      console.log(orderedPackages);
+
       expect(flatten(orderedPackages)).toStrictEqual(['addon-template']);
       expect(flatten(orderedPackages[0].content.modules.topSort())).toStrictEqual([
         'addon/components/greet.js',
